Move AttestationScoreCache to the optimizations module
It is a precompute cache wrapping precompute_chain_attestation_scores, not a
spec function, so it belongs alongside the other caches in optimizations.rs
(keeping lib.rs reading as the spec algorithm).
